Split world assembly economics
