+++ src/main/scala/com/boombustgroup/amorfati/engine/economics/OpenEconEconomics.scala
@@ -104,13 +104,13 @@ object OpenEconEconomics:
   case class StepInput(
       w: World,
       ledgerFinancialState: LedgerFinancialState,
-      s1: FiscalConstraintEconomics.Output,
-      s2: LaborEconomics.Output,
-      s3: HouseholdIncomeEconomics.Output,
-      s4: DemandEconomics.Output,
+      s1: FiscalConstraintEconomics.StepOutput,
+      s2: LaborEconomics.StepOutput,
+      s3: HouseholdIncomeEconomics.StepOutput,
+      s4: DemandEconomics.StepOutput,
       s5: FirmEconomics.StepOutput,
-      s6: HouseholdFinancialEconomics.Output,
-      s7: PriceEquityEconomics.Output,
+      s6: HouseholdFinancialEconomics.StepOutput,
+      s7: PriceEquityEconomics.StepOutput,
       banks: Vector[Banking.BankState],
       commodityRng: RandomStream,
   )
